pyproject.toml
[project]
name = "demo-book-recoo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.2.0",
    "notebook>=7.5.5",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "sentence-transformers>=5.3.0",
    "tabulate>=0.10.0",
]
